requirements.txt
typing-extensions>=4.0.0
regex>=2022.1.18
openai==1.12.0
python-dotenv==1.0.0 